backend/requirements.txt
fastapi
uvicorn
google-generativeai
pymupdf
reportlab
python-multipart
pydantic
